requirements.txt
langchain
openai==0.28.1
streamlit
python-dotenv
pydantic
faiss-gpu
unstructured
tiktoken
python-magic
redis
redisvl